requirements.txt
openai
python-dotenv
requests
hubspot-api-client
pandas
flask
plotly
jsonschema
pyyaml